post controller user validation in progress

+++ client/app/Services/PostService.js
@@ -1,19 +1,33 @@
 import { ProxyState } from "../AppState.js"
 import Post from "../Models/Post.js"
-// import { api } from "../Services/AxiosService.js"
 import { api } from "../Services/AxiosService.js"
 
-class PostService {
-  edit(id, editData) {
-    let res = api.put("posts/" + id, editData)
-    ProxyState.posts = ProxyState.posts
-  }
-  delete(id) {
-    let res = api.delete("posts/" + id)
-    ProxyState.posts = ProxyState.posts.filter(p => p.id != id)
-  }
+export default class PostService {
   constructor() {
     console.log("hello from post service")
+  }
+  async edit(id, editData) {
+    // validate current user is creator
+    let post = ProxyState.posts.find(p => p.id == id)
+    let userId = ProxyState.profile._id
+    if (userId != post.id) {
+      throw 'You are not the creator of this post.'
+    } else { 
+      let res = await api.put("posts/" + id, editData)
+      let postIndex = ProxyState.posts.findIndex(p => p.id == id)
+      ProxyState.posts = ProxyState.posts
+    }
+  }
+  async delete(id) {
+    let post = ProxyState.posts.find(p => p.id == id)
+    // validate current user is creator
+    let userId = ProxyState.profile._id
+    if (userId != post.id) {
+      throw 'You are not the creator of this post.'
+    } else { 
+      let res = await api.delete("posts/" + id)
+    }
+    ProxyState.posts = ProxyState.posts.filter(p => p.id != id)
   }
 
   async getPosts() {
@@ -30,5 +44,3 @@ class PostService {
 }
 
 export const postService = new PostService()
-
-export default postService;
